Stop using config.json, move everything to ENV
