+++ .example.env
@@ -21,6 +21,7 @@ DBL_TOKEN=
 # Time and language, you can leave it like this
 TIMEZONE=Europe/Paris
 DEFAULT_LANG=en
+SUPPORTED_LANGS=en fr ro nl ru sv zh de ja pt
 
 # Verbose mode allows me to get more information in case of bugs.
 # set it to 1 and let the bot run before sending me your logs
@@ -29,7 +30,9 @@ VERBOSE=0
 # App stuff
 PREFIX=!!
 BOT_NAME=QTweet
-
+MOD_ROLE=qtweet-mod
+PROFILE_URL=https://discordbots.org/bot/433615162394804224
+OWNER_ID=135712262156451840
 ## These variables set how fast the bot spawns shards. We avoid doing it too quickly.
 ## I don't recommend touching this, it won't affect you if your bot is in under 1000
 ## servers anyway.
