+++ Dockerfile
@@ -27,7 +27,6 @@ RUN yarn build
 FROM node:17.6-alpine
 WORKDIR /app
 
-COPY config.json .
 COPY package.json .
 COPY yarn.lock .
 RUN yarn install --production
@@ -37,4 +36,4 @@ COPY --from=langbuilder /lang/*.o.ftl lang/
 # Copy dist files over
 COPY --from=compiler /app/dist ./dist
 
-CMD [ "node", "-r", "esm", "dist/src/index.js" ]
+CMD [ "yarn", "start" ]
