--- config.json
@@ -1,6 +1,0 @@
-{
-  "ownerID": "135712262156451840",
-  "profileURL": "https://discordbots.org/bot/433615162394804224",
-  "modRole": "qtweet-mod",
-  "supportedLangs": ["en", "fr", "ro", "nl", "ru", "sv", "zh", "de", "ja", "pt"]
-}
