+++ docker-compose.yml
@@ -21,18 +21,22 @@ services:
       TWITTER_ACCESS_TOKEN: ${TWITTER_ACCESS_TOKEN}
       TWITTER_ACCESS_TOKEN_SECRET: ${TWITTER_ACCESS_TOKEN_SECRET}
       DBL_TOKEN: ${DBL_TOKEN}
-      PREFIX: ${PREFIX}
-      DEFAULT_LANG: ${DEFAULT_LANG}
-      BOT_NAME: ${BOT_NAME}
-      SHARD_SPAWN_DELAY: ${SHARD_SPAWN_DELAY}
-      SHARD_SPAWN_TIMEOUT: ${SHARD_SPAWN_TIMEOUT}
+      PREFIX: ${PREFIX:-!!}
+      DEFAULT_LANG: ${DEFAULT_LANG:-en}
+      BOT_NAME: ${BOT_NAME:-QTweet}
+      SHARD_SPAWN_DELAY: ${SHARD_SPAWN_DELAY:-20000}
+      SHARD_SPAWN_TIMEOUT: ${SHARD_SPAWN_TIMEOUT:-120000}
       VERBOSE: ${VERBOSE}
-      TWITTER_MAX_RECONNECT_DELAY: ${TWITTER_MAX_RECONNECT_DELAY}
-      TWEETS_TIMEOUT: ${TWEETS_TIMEOUT}
-      USERS_CHECK_TIMEOUT: ${USERS_CHECK_TIMEOUT}
-      USERS_BATCH_SIZE: ${USERS_BATCH_SIZE}
-      DISABLE_SANITY_CHECK: ${DISABLE_SANITY_CHECK}
-      DISABLE_STREAMS: ${DISABLE_STREAMS}
+      TWITTER_MAX_RECONNECT_DELAY: ${TWITTER_MAX_RECONNECT_DELAY:-240000}
+      TWEETS_TIMEOUT: ${TWEETS_TIMEOUT:-3600}
+      USERS_CHECK_TIMEOUT: ${USERS_CHECK_TIMEOUT:-1800}
+      USERS_BATCH_SIZE: ${USERS_BATCH_SIZE:-500}
+      DISABLE_SANITY_CHECK: ${DISABLE_SANITY_CHECK:-0}
+      DISABLE_STREAMS: ${DISABLE_STREAMS:-0}
+      MOD_ROLE: ${MOD_ROLE:-qtweet-mod}
+      PROFILE_URL: ${PROFILE_URL:-https://discordbots.org/bot/433615162394804224}
+      OWNER_ID: ${OWNER_ID:-135712262156451840}
+      SUPPORTED_LANGS: ${SUPPORTED_LANGS:-en fr ro nl ru sv zh de ja pt}
     depends_on:
       - postgres
