+++ src/db/index.ts
@@ -1,6 +1,8 @@
-import { Pool } from 'pg';
+import pg from 'pg';
+const { Pool } = pg;
 import log from '../log';
 
+type Pool = InstanceType<typeof Pool>
 export let pool: Pool;
 
 export const init = () => {
