+++ src/shard/commands/help.ts
@@ -1,16 +1,16 @@
 import { MessageEmbed } from "discord.js";
 import i18n from "../i18n";
-import { profileURL } from '../../../config.json';
 import { embed } from "../post";
 import { CmdFn } from ".";
 import { getGuildInfo } from "../../db/guilds";
+import process from 'process'
 
 const help : CmdFn = async (_, qChannel) => {
     const { lang: language, prefix } = await getGuildInfo(qChannel.guildId());
     const helpMsg = new MessageEmbed()
       .setColor(0x0e7675)
       .setTitle(i18n(language, 'helpHeader'))
-      .setURL(profileURL)
+      .setURL(process.env.PROFILE_URL)
       .setDescription(i18n(language, 'helpIntro'))
       .addField(`${prefix}tweet`, i18n(language, 'usage-tweet'))
       .addField(`${prefix}start`, i18n(language, 'usage-start'))
